static type-checking
Enable static type-checking. This is comprised of the following.

Add development dependencies:
- sorbet: type checker
- tapioca: generator of RBI files used in type checking

Note that we exclude sorbet-runtime because we're intentionally avoiding
making sorbet a runtime dependency of applications that use our gem.
This is also why we don't place any type annotations in the source code
of lib/schema_version_cache.rb itself, but rather in the rbi/ directory.
Files in that directory can be referenced by applications that use our
gem if they wish to perform type checking on their own codebase and its
use of our gem, but a project that doesn't do any type checking can
still use our gem.

Initialize the sorbet and tapioca configs.

Update bin/lint and bin/ci-test so that CI will:
- check for any out of date RBI files for the gems we depend on
- run the type checker

Add a `typed: true` annotation to lib/schema_version_cache.rb and add an
RBI file describing our public interface in the rbi/ directory.