don't use `git ls-files` in gemspec
Don't use `git ls-files` when constructing a file list in the gemspec.

This approach became widespread, but is now discouraged, because
building the gem from source will fail on a host that lacks git, and
bundling will in some circumstances emit an annoying error/warning:

    fatal: Not a git repository (or any of the parent directories): .git

Prior to this commit, the gemspec's `files` attribute resolved to:

    ["Gemfile", "README.md", "lib/schema_version_cache.rb", "rbi/schema_version_cache.rbi"]

I've additionally removed Gemfile from that list, since that's strictly
for use while developing the gem. The resulting `files` list is:

    ["README.md", "lib/schema_version_cache.rb", "rbi/schema_version_cache.rbi"]